feat: upgrade app to use next 14
