+++ .env.default
@@ -1,13 +1,12 @@
-NODE_ENVIRONMENT=development
+NODE_ENV=development
 DEBUG=*
-SERVER_PORT=3000
 LOG_LEVEL=debug
-ROOT_NODE_ID=johndoe@example.com
 WORKSPACE_CUSTOMER_ID=A0123b4c5
-WORKSPACE_DOMAIN=domin.com
+WORKSPACE_DOMAIN=domain.com
 AUTH_CLIENT_SUBJECT=admin@domain.com
 CACHE_ENGINE=memory
-REDIS_URL=redis://localhost:6380
-ADMIN_USERS_TTL_CACHE_S=3600
+CSV_FILE_PATH=./csv_sample.csv
 USER_BACKEND_MODE=csv
-CSV_FILE_PATH=csv_sample.csv
+REDIS_URL=redis://localhost:6380
+TTL_CACHE_S=3600
+ROOT_NODE_ID=johndoe@example.com
