+++ package.json
@@ -5,17 +5,19 @@
   "repository": "https://github.com/brunofurmon/open-org-chart",
   "author": "Bruno Furtado",
   "scripts": {
-    "lint": "eslint --fix ./**/*.js",
-    "debug": "nodemon --inspect=0.0.0.0 src/application/index.js",
+    "lint": "next lint --fix",
+    "dev": "next dev",
     "test": "jest --detectOpenHandles --coverage",
-    "start": "node src/application/index.js",
+    "start": "next start",
     "build": "next build",
-    "build-analyzer": "ANALYZE=true yarn build",
-    "postversion": "git push --follow-tags"
+    "build-analyzer": "ANALYZE=true next build",
+    "postversion": "git push --follow-tags",
+    "nextdev": "next dev"
   },
   "devDependencies": {
     "eslint": "^7.32.0 || ^8.2.0",
     "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-next": "14.2.8",
     "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jest": "^27.2.2",
     "jest": "^29.6.1",
@@ -35,7 +37,7 @@
     "googleapis": "105",
     "helmet": "^7.0.0",
     "http-status": "^1.6.2",
-    "next": "^13.4.10",
+    "next": "^14.2.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "redis": "^4.6.7",
