+++ src/app/api/(healthcheck)/liveness/route.js
@@ -1,0 +1,7 @@
+import { NextResponse } from 'next/server';
+
+export const GET = async (_) => NextResponse.json({
+  status: 'ok',
+  data: 'Server is up and running',
+  serverDate: new Date().toISOString(),
+});
