--- src/components/pages/Home/index.js
@@ -1,1 +1,0 @@
-export { default } from './Home';
