+++ src/domain/user/service.js
@@ -1,62 +1,67 @@
-const init = ({ fetchUsers }) => {
-  const listUsers = async (debugMode = false) => {
-    let users = await fetchUsers();
+import container from '@/src/container';
 
-    // Creates an extra node behind the root to aggregate parentless nodes, removing them from the top level
-    const fixmeIntermediateNode = {
-      name: "Fix Me",
-      id: "fixme@fixme.com",
-      email: "fixme@fixme.com",
-      managedByEmail: process.env.ROOT_NODE_ID,
-      parentId: process.env.ROOT_NODE_ID,
-      photoUrl: "http://placekitten.com/g/200/300",
-      title: "",
-      area: "",
-    };
-    users.push(fixmeIntermediateNode);
+const userBackendMode = process.env.USER_BACKEND_MODE || "csv";
 
-    const existingParentsList = users.map((user) =>
-      user.email.trim().toLowerCase()
-    );
+const usersStore =
+  userBackendMode === 'csv'
+    ? container.resolve('csvUserStore')
+    : container.resolve('googleadminUserStore');
+const { cachedResult } = container.resolve('cache');
 
-    for (let i = 0; i < users.length; i++) {
-      // Reassign parent id to root node, except for the root node
-      const user = users[i];
-      if (user.email.trim().toLowerCase() === process.env.ROOT_NODE_ID) {
-        user.id = user.email.trim().toLowerCase();
-        user.parentId = "";
-        continue;
-      }
+export const listUsers = async (debugMode = false) => {
+  let users = await cachedResult(
+    usersStore.listUsers,
+    'listUsers',
+    process.env.CACHE_TTL
+  );
 
-      if (
-        !existingParentsList.includes(user.managedByEmail?.trim().toLowerCase())
-      ) {
-        user.managedByEmail = fixmeIntermediateNode.id;
+  // Creates an extra node behind the root to aggregate parentless nodes, removing them from the top level
+  const fixmeIntermediateNode = {
+    name: "Fix Me",
+    id: "fixme@fixme.com",
+    email: "fixme@fixme.com",
+    managedByEmail: process.env.ROOT_NODE_ID,
+    parentId: process.env.ROOT_NODE_ID,
+    photoUrl: "https://placecats.com/g/200/200",
+    title: "",
+    area: "",
+  };
+  users.push(fixmeIntermediateNode);
 
-        // go down the tree and mark other users as managed by the intermediate node too
-        const managedUsers = users.filter(mUser => mUser.managedByEmail === user.email);
-        managedUsers.forEach(managedUser => {
-          managedUser.managedByEmail = fixmeIntermediateNode.id;
-        });
-      }
+  const existingParentsList = users.map((user) =>
+    user.email.trim().toLowerCase()
+  );
 
+  for (let i = 0; i < users.length; i++) {
+    // Reassign parent id to root node, except for the root node
+    const user = users[i];
+    if (user.email.trim().toLowerCase() === process.env.ROOT_NODE_ID) {
       user.id = user.email.trim().toLowerCase();
-      user.parentId = user.managedByEmail.trim().toLowerCase();
-    }
-    if (!debugMode) {
-      users = users.filter(
-        (user) =>
-          user.parentId !== fixmeIntermediateNode.id &&
-          user.id !== fixmeIntermediateNode.id
-      );
+      user.parentId = "";
+      continue;
     }
 
-    return users;
-  };
+    if (
+      !existingParentsList.includes(user.managedByEmail?.trim().toLowerCase())
+    ) {
+      user.managedByEmail = fixmeIntermediateNode.id;
 
-  return {
-    listUsers,
-  };
+      // go down the tree and mark other users as managed by the intermediate node too
+      const managedUsers = users.filter(mUser => mUser.managedByEmail === user.email);
+      managedUsers.forEach(managedUser => {
+        managedUser.managedByEmail = fixmeIntermediateNode.id;
+      });
+    }
+
+    user.id = user.email.trim().toLowerCase();
+    user.parentId = user.managedByEmail.trim().toLowerCase();
+  }
+  if (!debugMode) {
+    users = users.filter(
+      (user) =>
+        user.parentId !== fixmeIntermediateNode.id &&
+        user.id !== fixmeIntermediateNode.id
+    );
+  }
+  return users;
 };
-
-module.exports = init;
