--- src/integration/csv/csv_sample.csv
@@ -1,26 +1,0 @@
-Full Name,Email,Photo URL,Managed By Email,Title,Area
-John Doe,johndoe@example.com,https://placekitten.com/200/300, ,Software Engineer,Engineering
-Jane Smith,janesmith@example.com,https://placekitten.com/200/301,johndoe@example.com,Product Manager,Product
-Bob Johnson,bobjohnson@example.com,https://placekitten.com/200/302,johndoe@example.com,Designer,Design
-Alice Williams,alicewilliams@example.com,https://placekitten.com/200/303,johndoe@example.com,QA Engineer,Quality Assurance
-Charlie Brown,charliebrown@example.com,https://placekitten.com/200/304,janesmith@example.com,DevOps Engineer,Operations
-Diana Prince,dianaprince@example.com,https://placekitten.com/200/305,janesmith@example.com,HR Manager,Human Resources
-Edward Norton,edwardnorton@example.com,https://placekitten.com/200/306,janesmith@example.com,Marketing Specialist,Marketing
-Fiona Apple,fionaapple@example.com,https://placekitten.com/200/307,bobjohnson@example.com,Sales Executive,Sales
-George Clooney,georgeclooney@example.com,https://placekitten.com/200/308,bobjohnson@example.com,Support Engineer,Support
-Hannah Montana,hannahmontana@example.com,https://placekitten.com/200/309,bobjohnson@example.com,Finance Analyst,Finance
-Isaac Newton,isaacnewton@example.com,https://placekitten.com/200/310,alicewilliams@example.com,Research Scientist,Research
-Jack Sparrow,jacksparrow@example.com,https://placekitten.com/200/311,alicewilliams@example.com,Navigator,Navigation
-Katy Perry,katyperry@example.com,https://placekitten.com/200/312,alicewilliams@example.com,Singer,Entertainment
-Leonardo DiCaprio,leonardodicaprio@example.com,https://placekitten.com/200/313,charliebrown@example.com,Actor,Entertainment
-Mona Lisa,monalisa@example.com,https://placekitten.com/200/314,charliebrown@example.com,Art Curator,Art
-Neil Armstrong,neilarmstrong@example.com,https://placekitten.com/200/315,charliebrown@example.com,Astronaut,Aerospace
-Oprah Winfrey,oprahwinfrey@example.com,https://placekitten.com/200/316,dianaprince@example.com,TV Host,Media
-Pablo Picasso,pablopicasso@example.com,https://placekitten.com/200/317,dianaprince@example.com,Artist,Art
-Quentin Tarantino,quentintarantino@example.com,https://placekitten.com/200/318,dianaprince@example.com,Director,Film
-Rihanna,rihanna@example.com,https://placekitten.com/200/319,edwardnorton@example.com,Singer,Entertainment
-Steve Jobs,stevejobs@example.com,https://placekitten.com/200/320,edwardnorton@example.com,CEO,Technology
-Taylor Swift,taylorswift@example.com,https://placekitten.com/200/321,edwardnorton@example.com,Singer,Entertainment
-Usain Bolt,usainbolt@example.com,https://placekitten.com/200/322,fionaapple@example.com,Athlete,Sports
-Vincent van Gogh,vincentvangogh@example.com,https://placekitten.com/200/323,fionaapple@example.com,Artist,Art
-Walt Disney,waltdisney@example.com,https://placekitten.com/200/324,fionaapple@example.com,Animator,Entertainment
