--- src/pages/_document.js
@@ -1,15 +1,0 @@
-import { Html, Head, Main, NextScript } from "next/document";
-
-const Document = () => {
-  return (
-    <Html>
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-};
-
-export default Document;
