--- src/pages/index.js
@@ -1,3 +1,0 @@
-import Home from "../components/pages/Home";
-
-export default Home;
